Separate flex flow serving test
